refactor: apply separation of concerns
